add test that includes '\'\t' in blobField
This tests the LocalInfile string escape mechanism; during
LOAD DATA INFILE for data being received by the czar from a worker
via protobuf, we specify that the fields are enclosed by single
quotes, and accept the default that the fields are separated by
tabs. This test uses a blob that contains a single quote followed
by a tab inside the blob. This should not result in a spurious end
of the field because the tab should be escaped by LocalInfile so
that the chars passed into mysql from there are modified from
`'\t` to `'\\t` and so the sequence is not interpreted as the end
of the field.