disable test that can't ve verified with antlr 2
This query is not parsable by the older antlr 2 parser code and was
broken in this ticket in the antlr4 parser code. The qserv manual says that
BETWEEN in a WHERE clause is not supported *and* functions in the
ORDER_BY clause are also not supported. So I'm disabling this for now. I
created ticket DM-16237 to re-add support for this in the anltr4 parser
and create a unit test for this ticket that does not rely on the antlr 2
parser for verification.